fix: strip all non-syntactic spaces from cpp/java snippets

+++ lib/snippets.ts
@@ -94,7 +94,7 @@ export const SNIPPETS: Snippet[] = [
     language: "java",
     url: "https://leetcode.com/problems/binary-search/",
     code: `int search(int[] nums, int target){
-    int lo=0, hi=nums.length-1;
+    int lo=0,hi=nums.length-1;
     while(lo<=hi){
         int mid=lo+(hi-lo)/2;
         if(nums[mid]==target) return mid;
@@ -111,8 +111,8 @@ export const SNIPPETS: Snippet[] = [
     language: "cpp",
     url: "https://leetcode.com/problems/climbing-stairs/",
     code: `int climb_stairs(int n){
-    int a=1, b=1;
-    for(int i=0; i<n; i++){
+    int a=1,b=1;
+    for(int i=0;i<n;i++){
         int t=a+b;
         a=b;
         b=t;
@@ -659,7 +659,7 @@ def max_sliding_window(nums, k):
     difficulty: "easy",
     language: "cpp",
     code: `int binary_search(vector<int>& nums, int target){
-    int lo=0, hi=nums.size()-1;
+    int lo=0,hi=nums.size()-1;
     while(lo<=hi){
         int mid=lo+(hi-lo)/2;
         if(nums[mid]==target) return mid;
@@ -676,9 +676,9 @@ def max_sliding_window(nums, k):
     language: "cpp",
     code: `vector<int> two_sum(vector<int>& nums, int target){
     unordered_map<int,int> seen;
-    for(int i=0; i<nums.size(); i++){
+    for(int i=0;i<(int)nums.size();i++){
         int comp=target-nums[i];
-        if(seen.count(comp)) return {seen[comp], i};
+        if(seen.count(comp)) return {seen[comp],i};
         seen[nums[i]]=i;
     }
     return {};
@@ -690,10 +690,10 @@ def max_sliding_window(nums, k):
     difficulty: "medium",
     language: "cpp",
     code: `int max_subarray(vector<int>& nums){
-    int best=nums[0], cur=nums[0];
-    for(int i=1; i<nums.size(); i++){
-        cur=max(nums[i], cur+nums[i]);
-        best=max(best, cur);
+    int best=nums[0],cur=nums[0];
+    for(int i=1;i<(int)nums.size();i++){
+        cur=max(nums[i],cur+nums[i]);
+        best=max(best,cur);
     }
     return best;
 }`,
@@ -705,13 +705,13 @@ def max_sliding_window(nums, k):
     language: "cpp",
     code: `bool is_valid(string s){
     stack<char> st;
-    for(char c : s){
-        if(c=='(' || c=='[' || c=='{') st.push(c);
+    for(char c:s){
+        if(c=='('||c=='['||c=='{') st.push(c);
         else{
             if(st.empty()) return false;
-            if(c==')' && st.top()!='(') return false;
-            if(c==']' && st.top()!='[') return false;
-            if(c=='}' && st.top()!='{') return false;
+            if(c==')'&&st.top()!='(') return false;
+            if(c==']'&&st.top()!='[') return false;
+            if(c=='}'&&st.top()!='{') return false;
             st.pop();
         }
     }
@@ -740,13 +740,13 @@ def max_sliding_window(nums, k):
     difficulty: "medium",
     language: "cpp",
     code: `vector<vector<int>> merge(vector<vector<int>>& intervals){
-    sort(intervals.begin(), intervals.end());
+    sort(intervals.begin(),intervals.end());
     vector<vector<int>> res;
-    for(auto& iv : intervals){
-        if(res.empty() || iv[0]>res.back()[1])
+    for(auto& iv:intervals){
+        if(res.empty()||iv[0]>res.back()[1])
             res.push_back(iv);
         else
-            res.back()[1]=max(res.back()[1], iv[1]);
+            res.back()[1]=max(res.back()[1],iv[1]);
     }
     return res;
 }`,
